+++ src/aks-preview/azext_aks_preview/tests/latest/test_managed_cluster_decorator.py
@@ -1116,9 +1116,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(
-            ctx_1.get_acns_enablement_with_perf(), (None, None, None, None)
-        )
+        self.assertEqual(ctx_1.get_acns_enablement(), (None, None, None, None))
 
         # Flag set to True.
         ctx_2 = AKSPreviewManagedClusterContext(
@@ -1131,9 +1129,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(
-            ctx_2.get_acns_enablement_with_perf(), (True, None, None, None)
-        )
+        self.assertEqual(ctx_2.get_acns_enablement(), (True, None, None, None))
 
         # Flag set to True.
         ctx_3 = AKSPreviewManagedClusterContext(
@@ -1146,9 +1142,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(
-            ctx_3.get_acns_enablement_with_perf(), (True, None, None, None)
-        )
+        self.assertEqual(ctx_3.get_acns_enablement(), (True, None, None, None))
 
         # Flag set to True and False.
         ctx_4 = AKSPreviewManagedClusterContext(
@@ -1164,7 +1158,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
         )
         # fail on get_acns mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_4.get_acns_enablement_with_perf()
+            ctx_4.get_acns_enablement()
 
         # Flag set to False.
         ctx_5 = AKSPreviewManagedClusterContext(
@@ -1177,9 +1171,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(
-            ctx_5.get_acns_enablement_with_perf(), (False, None, None, None)
-        )
+        self.assertEqual(ctx_5.get_acns_enablement(), (False, None, None, None))
 
         ctx_6 = AKSPreviewManagedClusterContext(
             self.cmd,
@@ -1192,9 +1184,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(
-            ctx_6.get_acns_enablement_with_perf(), (True, False, None, None)
-        )
+        self.assertEqual(ctx_6.get_acns_enablement(), (True, False, None, None))
 
         ctx_7 = AKSPreviewManagedClusterContext(
             self.cmd,
@@ -1207,9 +1197,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(
-            ctx_7.get_acns_enablement_with_perf(), (True, None, False, None)
-        )
+        self.assertEqual(ctx_7.get_acns_enablement(), (True, None, False, None))
 
         # Cannot disable observability with enabling acns
         ctx_8 = AKSPreviewManagedClusterContext(
@@ -1222,9 +1210,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(
-            ctx_8.get_acns_enablement_with_perf(), (None, None, None, None)
-        )
+        self.assertEqual(ctx_8.get_acns_enablement(), (None, None, None, None))
 
         # Cannot disable security with enabling acns
         ctx_9 = AKSPreviewManagedClusterContext(
@@ -1237,9 +1223,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(
-            ctx_9.get_acns_enablement_with_perf(), (None, None, None, None)
-        )
+        self.assertEqual(ctx_9.get_acns_enablement(), (None, None, None, None))
 
         # Illegal flags enable acns, disable acns security, disable acns observability
         ctx_10 = AKSPreviewManagedClusterContext(
@@ -1254,9 +1238,9 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        # fail on get_acns_enablement_with_perf mutual exclusive error
+        # fail on get_acns_enablement mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_10.get_acns_enablement_with_perf()
+            ctx_10.get_acns_enablement()
 
         # Illegal flags disable acns and disable acns observability
         ctx_11 = AKSPreviewManagedClusterContext(
@@ -1270,9 +1254,9 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        # fail on get_acns_enablement_with_perf mutual exclusive error
+        # fail on get_acns_enablement mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_11.get_acns_enablement_with_perf()
+            ctx_11.get_acns_enablement()
 
         # Illegal flags disable acns and disable acns security
         ctx_12 = AKSPreviewManagedClusterContext(
@@ -1286,9 +1270,9 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        # fail on get_acns_enablement_with_perf mutual exclusive error
+        # fail on get_acns_enablement mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_12.get_acns_enablement_with_perf()
+            ctx_12.get_acns_enablement()
 
         # Enable ACNS and ACNS performance
         ctx_13 = AKSPreviewManagedClusterContext(
@@ -1302,9 +1286,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(
-            ctx_13.get_acns_enablement_with_perf(), (True, None, None, True)
-        )
+        self.assertEqual(ctx_13.get_acns_enablement(), (True, None, None, True))
 
         # Enable ACNS and ACNS performance with disable acns security, acns observability
         ctx_14 = AKSPreviewManagedClusterContext(
@@ -1320,9 +1302,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        self.assertEqual(
-            ctx_14.get_acns_enablement_with_perf(), (True, False, False, True)
-        )
+        self.assertEqual(ctx_14.get_acns_enablement(), (True, False, False, True))
 
         # Enable all of ACNS (security and observability unspecified)
         ctx_15 = AKSPreviewManagedClusterContext(
@@ -1336,9 +1316,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(
-            ctx_15.get_acns_enablement_with_perf(), (True, None, None, True)
-        )
+        self.assertEqual(ctx_15.get_acns_enablement(), (True, None, None, True))
 
         # Enable ACNS, disable performance, security and observability unspecified
         ctx_15 = AKSPreviewManagedClusterContext(
@@ -1352,9 +1330,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        self.assertEqual(
-            ctx_15.get_acns_enablement_with_perf(), (True, None, None, False)
-        )
+        self.assertEqual(ctx_15.get_acns_enablement(), (True, None, None, False))
 
         # Illegal flags disable acns and disable acns performance
         ctx_16 = AKSPreviewManagedClusterContext(
@@ -1368,9 +1344,9 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.UPDATE,
         )
-        # fail on get_acns_enablement_with_perf mutual exclusive error
+        # fail on get_acns_enablement mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_16.get_acns_enablement_with_perf()
+            ctx_16.get_acns_enablement()
 
         # Illegal flags enable acns and all suites disabled
         ctx_17 = AKSPreviewManagedClusterContext(
@@ -1386,9 +1362,9 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
             self.models,
             decorator_mode=DecoratorMode.CREATE,
         )
-        # fail on get_acns_enablement_with_perf mutual exclusive error
+        # fail on get_acns_enablement mutual exclusive error
         with self.assertRaises(MutuallyExclusiveArgumentError):
-            ctx_17.get_acns_enablement_with_perf()
+            ctx_17.get_acns_enablement()
 
     def test_get_enable_managed_identity(self):
         # custom value